+++ includes/cleanup-head.php
@@ -1,8 +1,7 @@
 <?php
 
-// if ( current_theme_supports( 'ejo-cleanup-frontend', 'head' ) ) {
+if ( current_theme_supports( 'ejo-cleanup-frontend', 'head' ) ) {
 	remove_action( 'wp_head', 'wlwmanifest_link' ); //* Remove unnecessary Window Live Writer link
 	remove_action( 'wp_head', 'hybrid_meta_generator', 1 ); //* Remove unnecessary Theme version
-// }
-
-	write_log( 'generator weghalen!');
+	remove_action( 'wp_head', 'wp_generator' ); //* Remove Wordpress version
+}