+++ includes/disable-pingback.php
@@ -1,28 +1,30 @@
 <?php 
 
-//* Prevent pingback being sent to XMLRPC server
-//* (Probably not necessary if XML-RPC disabled)
-add_filter( 'xmlrpc_methods', 'ejo_prevent_pingback_xmlrpc' );
+if ( current_theme_supports( 'ejo-cleanup-frontend', 'pingback' ) ) {
 
-//* Remove unnecesary HTTP Header response item
-add_filter( 'wp_headers', 'ejo_remove_x_pingback_header' );
+	//* Prevent pingback being sent to XMLRPC server
+	//* (Probably not necessary if XML-RPC disabled)
+	add_filter( 'xmlrpc_methods', 'ejo_prevent_pingback_xmlrpc' );
 
-//* Remove unnecesary <head> xmlrpc element (not related to pingback I think)
-remove_action( 'wp_head', 'rsd_link' );
-remove_action( 'wp_head', 'hybrid_link_pingback',  3 );
+	//* Remove unnecesary HTTP Header response item
+	add_filter( 'wp_headers', 'ejo_remove_x_pingback_header' );
+
+	//* Remove Hybrid pingback link
+	remove_action( 'wp_head', 'hybrid_link_pingback',  3 );
 
 
-//* Prevent pingback being sent to XMLRPC server
-function ejo_prevent_pingback_xmlrpc( $methods ) 
-{
-	unset( $methods['pingback.ping'] );
-	unset( $methods['pingback.extensions.getPingbacks'] );
+	//* Prevent pingback being sent to XMLRPC server
+	function ejo_prevent_pingback_xmlrpc( $methods ) 
+	{
+		unset( $methods['pingback.ping'] );
+		unset( $methods['pingback.extensions.getPingbacks'] );
 
-	return $methods;
-}
+		return $methods;
+	}
 
-//* Remove from HTTP Header response
-function ejo_remove_x_pingback_header( $headers ) 
-{
-	unset( $headers['X-Pingback'] );
+	//* Remove from HTTP Header response
+	function ejo_remove_x_pingback_header( $headers ) 
+	{
+		unset( $headers['X-Pingback'] );
+	}
 }