+++ includes/helpers/theme_support_arguments.php
@@ -1,14 +1,4 @@
-<?php 
-//* Delete one array-record based on value
-function remove_array_value( $value, $array ) 
-{
-	$key = array_search($value, $array);
-
-	if( $key !== false)
-		unset($array[$key]);
-
-	return $array;
-}
+<?php
 
 /**
  * Wordpress only accepts arguments at 'current_theme_supports' function for